+++ src/main/java/nextstep/auth/AuthService.java
@@ -2,7 +2,7 @@ package nextstep.auth;
 
 import nextstep.member.Member;
 import nextstep.member.MemberDao;
-import nextstep.support.InvalidLoginException;
+import nextstep.exception.InvalidLoginException;
 import org.springframework.stereotype.Service;
 
 @Service
