+++ src/main/java/nextstep/exception/InvalidLoginException.java
@@ -1,4 +1,4 @@
-package nextstep.support;
+package nextstep.exception;
 
 public class InvalidLoginException extends RuntimeException {
