+++ src/main/java/nextstep/exception/UnauthorizedException.java
@@ -1,4 +1,4 @@
-package nextstep.support;
+package nextstep.exception;
 
 public class UnauthorizedException extends RuntimeException {
